fixed references, linked code to uitbouw, added namespaces, relocated scripts to correct folders
