+++ 3DAmsterdam/Assets/Netherlands3D/Scripts/Sharing/SceneSerializer.cs
@@ -3,6 +3,7 @@ using Netherlands3D.Cameras;
 using Netherlands3D.Interface;
 using Netherlands3D.Interface.Layers;
 using Netherlands3D.ObjectInteraction;
+using Netherlands3D.T3D.Uitbouw;
 using Newtonsoft.Json;
 using System;
 using System.Collections;
