+++ 3DAmsterdam/Assets/Netherlands3D/Scripts/Utilities/GeometryCalculator.cs
@@ -1,19 +1,23 @@
 using UnityEngine;
 
-public static class GeometryCalculator
+
+namespace Netherlands3D.Utilities
 {
-    public static bool ContainsPoint(Vector2[] polyPoints, Vector2 p)
+    public static class GeometryCalculator
     {
-        var j = polyPoints.Length - 1;
-        var inside = false;
-        for (int i = 0; i < polyPoints.Length; j = i++)
+        public static bool ContainsPoint(Vector2[] polyPoints, Vector2 p)
         {
-            var pi = polyPoints[i];
-            var pj = polyPoints[j];
-            if (((pi.y <= p.y && p.y < pj.y) || (pj.y <= p.y && p.y < pi.y)) &&
-                (p.x < (pj.x - pi.x) * (p.y - pi.y) / (pj.y - pi.y) + pi.x))
-                inside = !inside;
+            var j = polyPoints.Length - 1;
+            var inside = false;
+            for (int i = 0; i < polyPoints.Length; j = i++)
+            {
+                var pi = polyPoints[i];
+                var pj = polyPoints[j];
+                if (((pi.y <= p.y && p.y < pj.y) || (pj.y <= p.y && p.y < pi.y)) &&
+                    (p.x < (pj.x - pi.x) * (p.y - pi.y) / (pj.y - pi.y) + pi.x))
+                    inside = !inside;
+            }
+            return inside;
         }
-        return inside;
     }
 }
